# backend/static/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"OrderDetail": {
				"properties": {
					"Discount": {
						"type": "number",
						"format": "double"
					},
					"ProductId": {
						"type": "string"
					},
					"Quantity": {
						"type": "number",
						"format": "double"
					},
					"UnitPrice": {
						"type": "string"
					}
				},
				"required": [
					"Discount",
					"ProductId",
					"Quantity",
					"UnitPrice"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Orders": {
				"properties": {
					"CustomerUserName": {
						"type": "string"
					},
					"EmployeeUserName": {
						"type": "string"
					},
					"OrderDate": {
						"type": "string"
					},
					"TotalPrice": {
						"type": "number",
						"format": "double"
					},
					"State": {
						"type": "string"
					},
					"TableId": {
						"type": "string"
					},
					"TypeService": {
						"type": "string"
					},
					"OrderDetail": {
						"items": {
							"$ref": "#/components/schemas/OrderDetail"
						},
						"type": "array"
					}
				},
				"required": [
					"CustomerUserName",
					"EmployeeUserName",
					"OrderDate",
					"TotalPrice",
					"State",
					"OrderDetail"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OrderDetailUpdate": {
				"properties": {
					"Discount": {
						"type": "number",
						"format": "double"
					},
					"ProductId": {
						"type": "string"
					},
					"Quantity": {
						"type": "number",
						"format": "double"
					},
					"UnitPrice": {
						"type": "string"
					},
					"Id": {
						"type": "string"
					},
					"OrderId": {
						"type": "string"
					}
				},
				"required": [
					"Discount",
					"ProductId",
					"Quantity",
					"UnitPrice"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OrdersUpdate": {
				"properties": {
					"CustomerId": {
						"type": "string"
					},
					"EmployeeId": {
						"type": "string"
					},
					"OrderDate": {
						"type": "string"
					},
					"TotalPrice": {
						"type": "number",
						"format": "double"
					},
					"State": {
						"type": "string"
					},
					"TableId": {
						"type": "string"
					},
					"TypeService": {
						"type": "string"
					},
					"OrderDetail": {
						"items": {
							"$ref": "#/components/schemas/OrderDetailUpdate"
						},
						"type": "array"
					}
				},
				"required": [
					"CustomerId",
					"EmployeeId",
					"OrderDate",
					"TotalPrice",
					"State",
					"OrderDetail"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"keycloak": {
				"type": "apiKey",
				"name": "Authorization",
				"description": "Keycloak Bearer Token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/Order": {
			"get": {
				"operationId": "GetOrderAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"allOf": [
											{
												"properties": {
													"OrderDetail": {
														"items": {
															"properties": {
																"OrderId": {
																	"type": "string"
																},
																"Quantity": {
																	"type": "number",
																	"format": "double"
																},
																"ProductId": {
																	"type": "string"
																},
																"Discount": {
																	"type": "number",
																	"format": "double"
																},
																"UnitPrice": {
																	"type": "string"
																},
																"Id": {
																	"type": "string"
																}
															},
															"required": [
																"OrderId",
																"Quantity",
																"ProductId",
																"Discount",
																"UnitPrice",
																"Id"
															],
															"type": "object"
														},
														"type": "array"
													},
													"Employee": {
														"properties": {
															"PhotoPath": {
																"type": "string"
															},
															"BirthDate": {
																"type": "string"
															},
															"Photo": {
																"type": "string"
															},
															"Notes": {
																"type": "string"
															},
															"PostalCode": {
																"type": "string"
															},
															"Country": {
																"type": "string"
															},
															"City": {
																"type": "string"
															},
															"Address": {
																"type": "string"
															},
															"UserName": {
																"type": "string"
															},
															"Email": {
																"type": "string"
															},
															"LastName": {
																"type": "string"
															},
															"FirstName": {
																"type": "string"
															},
															"Title": {
																"type": "string"
															},
															"State": {
																"type": "string"
															},
															"Id": {
																"type": "string"
															}
														},
														"required": [
															"PhotoPath",
															"BirthDate",
															"Photo",
															"Notes",
															"PostalCode",
															"Country",
															"City",
															"Address",
															"UserName",
															"Email",
															"LastName",
															"FirstName",
															"Title",
															"State",
															"Id"
														],
														"type": "object"
													},
													"Customer": {
														"properties": {
															"Role": {
																"type": "string"
															},
															"PhotoPath": {
																"type": "string"
															},
															"BirthDate": {
																"type": "string"
															},
															"Photo": {
																"type": "string"
															},
															"Notes": {
																"type": "string"
															},
															"PostalCode": {
																"type": "string"
															},
															"Country": {
																"type": "string"
															},
															"City": {
																"type": "string"
															},
															"Address": {
																"type": "string"
															},
															"UserName": {
																"type": "string"
															},
															"Email": {
																"type": "string"
															},
															"LastName": {
																"type": "string"
															},
															"FirstName": {
																"type": "string"
															},
															"Title": {
																"type": "string"
															},
															"Id": {
																"type": "string"
															}
														},
														"required": [
															"Role",
															"PhotoPath",
															"BirthDate",
															"Photo",
															"Notes",
															"PostalCode",
															"Country",
															"City",
															"Address",
															"UserName",
															"Email",
															"LastName",
															"FirstName",
															"Title",
															"Id"
														],
														"type": "object"
													},
													"Table": {
														"properties": {
															"State": {
																"type": "string"
															},
															"Description": {
																"type": "string"
															},
															"Name": {
																"type": "string"
															},
															"Id": {
																"type": "string"
															}
														},
														"required": [
															"State",
															"Description",
															"Name",
															"Id"
														],
														"type": "object"
													}
												},
												"required": [
													"OrderDetail",
													"Employee",
													"Customer",
													"Table"
												],
												"type": "object"
											},
											{
												"properties": {
													"TableId": {
														"type": "string"
													},
													"TypeService": {
														"type": "string"
													},
													"TotalPrice": {
														"type": "number",
														"format": "double"
													},
													"OrderDate": {
														"type": "string"
													},
													"EmployeeId": {
														"type": "string"
													},
													"CustomerId": {
														"type": "string"
													},
													"State": {
														"type": "string"
													},
													"Id": {
														"type": "string"
													}
												},
												"required": [
													"TableId",
													"TypeService",
													"TotalPrice",
													"OrderDate",
													"EmployeeId",
													"CustomerId",
													"State",
													"Id"
												],
												"type": "object"
											}
										]
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateOrder",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"error": {},
												"message": {},
												"status": {},
												"orderDetails": {
													"items": {
														"properties": {
															"OrderId": {
																"type": "string"
															},
															"Quantity": {
																"type": "number",
																"format": "double"
															},
															"ProductId": {
																"type": "string"
															},
															"Discount": {
																"type": "number",
																"format": "double"
															},
															"UnitPrice": {
																"type": "string"
															},
															"Id": {
																"type": "string"
															}
														},
														"required": [
															"OrderId",
															"Quantity",
															"ProductId",
															"Discount",
															"UnitPrice",
															"Id"
														],
														"type": "object"
													},
													"type": "array"
												},
												"order": {
													"properties": {
														"TableId": {
															"type": "string"
														},
														"TypeService": {
															"type": "string"
														},
														"TotalPrice": {
															"type": "number",
															"format": "double"
														},
														"OrderDate": {
															"type": "string"
														},
														"EmployeeId": {
															"type": "string"
														},
														"CustomerId": {
															"type": "string"
														},
														"State": {
															"type": "string"
														},
														"Id": {
															"type": "string"
														}
													},
													"required": [
														"TableId",
														"TypeService",
														"TotalPrice",
														"OrderDate",
														"EmployeeId",
														"CustomerId",
														"State",
														"Id"
													],
													"type": "object"
												}
											},
											"required": [
												"orderDetails",
												"order"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {},
												"orderDetails": {},
												"order": {},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"message",
												"status"
											],
											"type": "object"
										},
										{
											"properties": {
												"orderDetails": {},
												"order": {},
												"error": {},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"error",
												"message",
												"status"
											],
											"type": "object"
										},
										{
											"type": "string",
											"enum": [
												"ผู้ใช้งานนี้ไม่สามารถเพิ่มข้อมูลได้",
												"ไม่พบข้อมูลลูกค้า",
												"ไม่พบข้อมูลพนักงาน"
											]
										}
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"keycloak": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Orders"
							}
						}
					}
				}
			}
		},
		"/Order/{id}": {
			"get": {
				"operationId": "GetOrderById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"allOf": [
												{
													"properties": {
														"OrderDetail": {
															"items": {
																"properties": {
																	"OrderId": {
																		"type": "string"
																	},
																	"Quantity": {
																		"type": "number",
																		"format": "double"
																	},
																	"ProductId": {
																		"type": "string"
																	},
																	"Discount": {
																		"type": "number",
																		"format": "double"
																	},
																	"UnitPrice": {
																		"type": "string"
																	},
																	"Id": {
																		"type": "string"
																	}
																},
																"required": [
																	"OrderId",
																	"Quantity",
																	"ProductId",
																	"Discount",
																	"UnitPrice",
																	"Id"
																],
																"type": "object"
															},
															"type": "array"
														},
														"Table": {
															"properties": {
																"State": {
																	"type": "string"
																},
																"Description": {
																	"type": "string"
																},
																"Name": {
																	"type": "string"
																},
																"Id": {
																	"type": "string"
																}
															},
															"required": [
																"State",
																"Description",
																"Name",
																"Id"
															],
															"type": "object"
														}
													},
													"required": [
														"OrderDetail",
														"Table"
													],
													"type": "object"
												},
												{
													"properties": {
														"TableId": {
															"type": "string"
														},
														"TypeService": {
															"type": "string"
														},
														"TotalPrice": {
															"type": "number",
															"format": "double"
														},
														"OrderDate": {
															"type": "string"
														},
														"EmployeeId": {
															"type": "string"
														},
														"CustomerId": {
															"type": "string"
														},
														"State": {
															"type": "string"
														},
														"Id": {
															"type": "string"
														}
													},
													"required": [
														"TableId",
														"TypeService",
														"TotalPrice",
														"OrderDate",
														"EmployeeId",
														"CustomerId",
														"State",
														"Id"
													],
													"type": "object"
												}
											]
										},
										{
											"type": "string",
											"enum": [
												"ไม่พบข้อมูล"
											]
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateOrder",
				"responses": {
					"200": {
						"description": "Update",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"error": {},
												"message": {},
												"status": {},
												"orderDetails": {
													"items": {
														"properties": {
															"OrderId": {
																"type": "string"
															},
															"Quantity": {
																"type": "number",
																"format": "double"
															},
															"ProductId": {
																"type": "string"
															},
															"Discount": {
																"type": "number",
																"format": "double"
															},
															"UnitPrice": {
																"type": "string"
															},
															"Id": {
																"type": "string"
															}
														},
														"required": [
															"OrderId",
															"Quantity",
															"ProductId",
															"Discount",
															"UnitPrice",
															"Id"
														],
														"type": "object"
													},
													"type": "array"
												},
												"order": {
													"properties": {
														"TableId": {
															"type": "string"
														},
														"TypeService": {
															"type": "string"
														},
														"TotalPrice": {
															"type": "number",
															"format": "double"
														},
														"OrderDate": {
															"type": "string"
														},
														"EmployeeId": {
															"type": "string"
														},
														"CustomerId": {
															"type": "string"
														},
														"State": {
															"type": "string"
														},
														"Id": {
															"type": "string"
														}
													},
													"required": [
														"TableId",
														"TypeService",
														"TotalPrice",
														"OrderDate",
														"EmployeeId",
														"CustomerId",
														"State",
														"Id"
													],
													"type": "object"
												}
											},
											"required": [
												"orderDetails",
												"order"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {},
												"orderDetails": {},
												"order": {},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"message",
												"status"
											],
											"type": "object"
										},
										{
											"properties": {
												"orderDetails": {},
												"order": {},
												"error": {},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"error",
												"message",
												"status"
											],
											"type": "object"
										},
										{
											"type": "string",
											"enum": [
												"ไม่พบข้อมูล",
												"ไม่พบข้อมูลลูกค้า",
												"ไม่พบข้อมูลพนักงาน"
											]
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/OrdersUpdate"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}